requirements.txt
discord.py==2.3.2
SQLAlchemy==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
openai==1.30.1
emoji==2.10.0
nltk==3.8.1
